chore: attempt externalize shiki
