+++ apps/web/package.json
@@ -51,6 +51,7 @@
         "polished": "^4.3.1",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
+        "shiki": "^1.6.5",
         "yjs": "^13.6.15",
         "zod": "^3.23.8"
     },
