+++ apps/web/src/components/HomeIntroStep4/HomeIntroStep4Code.tsx
@@ -2,7 +2,7 @@ import { Card } from "@pluv-internal/react-components/either";
 import { cn } from "@pluv-internal/utils";
 import { codeBlock } from "common-tags";
 import type { FC } from "react";
-import { ServerCodeBlock } from "../ServerCodeBlock";
+import { TypeScriptServerCodeBlock } from "../TypeScriptServerCodeBlock";
 
 export interface HomeIntroStep4CodeProps {
     className?: string;
@@ -11,7 +11,7 @@ export interface HomeIntroStep4CodeProps {
 export const HomeIntroStep4Code: FC<HomeIntroStep4CodeProps> = ({ className }) => {
     return (
         <Card className={cn("flex flex-col items-stretch overflow-auto shadow-md", className)}>
-            <ServerCodeBlock
+            <TypeScriptServerCodeBlock
                 className="grow"
                 code={codeBlock`
                     import { yjs } from "@pluv/crdt-yjs";
