+++ apps/web/src/components/MdxPre/index.ts
@@ -1,0 +1,1 @@
+export * from "./MdxPre";
