+++ apps/web/src/components/ServerCodeBlock/ServerCodeBlock.tsx
@@ -1,25 +1,16 @@
-import type { ShikiLanguage } from "@pluv-internal/react-components/utils";
-import { getShiki } from "@pluv-internal/react-components/utils";
 import type { InferComponentProps } from "@pluv-internal/typings";
 import { cn } from "@pluv-internal/utils";
-import type { TransformerTwoslashOptions } from "@shikijs/twoslash";
-import { rendererRich, transformerTwoslash } from "@shikijs/twoslash";
 import { oneLine } from "common-tags";
 import type { FC } from "react";
+import type { ShikiLanguage } from "../../utils/getShiki";
+import { getShiki } from "../../utils/getShiki";
 
 export type ServerCodeBlockProps = InferComponentProps<"div"> & {
     code: string;
     lang: ShikiLanguage;
-    twoslashOptions?: TransformerTwoslashOptions["twoslashOptions"];
 };
 
-export const ServerCodeBlock: FC<ServerCodeBlockProps> = async ({
-    className,
-    lang,
-    code,
-    twoslashOptions,
-    ...restProps
-}) => {
+export const ServerCodeBlock: FC<ServerCodeBlockProps> = async ({ className, lang, code, ...restProps }) => {
     const highlighter = await getShiki();
     const html = highlighter.codeToHtml(code, {
         lang,
@@ -27,23 +18,6 @@ export const ServerCodeBlock: FC<ServerCodeBlockProps> = async ({
             light: "catppuccin-latte",
             dark: "catppuccin-macchiato",
         },
-        transformers: [
-            transformerTwoslash({
-                renderer: rendererRich(),
-                twoslashOptions: {
-                    ...twoslashOptions,
-                    compilerOptions: {
-                        paths: {
-                            "@pluv/crdt-yjs": ["../../packages/crdt-yjs/dist"],
-                            "@pluv/io": ["../../packages/io/dist"],
-                            "@pluv/platform-node": ["../../packages/platform-node/dist"],
-                            "@pluv/react": ["../../packages/react/dist"],
-                        },
-                        ...twoslashOptions?.compilerOptions,
-                    },
-                },
-            }),
-        ],
     });
 
     return (
