+++ apps/web/src/components/TypeScriptServerCodeBlock/index.ts
@@ -1,0 +1,1 @@
+export * from "./TypeScriptServerCodeBlock";
