+++ internal/mdx-components/src/index.ts
@@ -24,10 +24,7 @@ export { MdxOl } from "./MdxOl";
 export type { MdxOlProps } from "./MdxOl";
 export { MdxP } from "./MdxP";
 export type { MdxPProps } from "./MdxP";
-export { MdxPre } from "./MdxPre";
-export type { MdxPreProps } from "./MdxPre";
 export { MdxTable } from "./MdxTable";
 export type { MdxTableProps } from "./MdxTable";
 export { MdxUl } from "./MdxUl";
 export type { MdxUlProps } from "./MdxUl";
-export { components } from "./components";
