+++ internal/react-components/package.json
@@ -11,7 +11,7 @@
 		"types.ts"
 	],
 	"scripts": {
-		"lint": "eslint {client,either,higher-order,server,utils}/**/*.ts* --fix",
+		"lint": "eslint {client,either,higher-order}/**/*.ts* --fix",
 		"tsc:check": "tsc --noEmit"
 	},
 	"publishConfig": {
@@ -56,7 +56,6 @@
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-hook-form": "^7.51.5",
-		"shiki": "^1.6.5",
 		"sonner": "^1.4.41"
 	},
 	"devDependencies": {
