--- internal/react-components/server/atoms/ServerCodeBlock/index.ts
@@ -1,1 +1,0 @@
-export * from "./ServerCodeBlock";
