--- internal/react-components/server/atoms/index.ts
@@ -1,1 +1,0 @@
-export * from "./ServerCodeBlock";
