--- internal/react-components/server/index.ts
@@ -1,1 +1,0 @@
-export * from "./atoms";
