--- internal/react-components/types.ts
@@ -1,1 +1,0 @@
-export type { ShikiLanguage, ShikiTheme } from "./utils/getShiki";
