--- internal/react-components/utils/index.ts
@@ -1,2 +1,0 @@
-export { getShiki } from "./getShiki";
-export type { ShikiLanguage } from "./getShiki";
